implemented the algorithm subcommand & Display for each config struct + added an id to BrokerApi
Just implemented the next part of the argument parser. It's now possible to list apis and algorithms. Also you can add new brokers. The individual accounts are distinguished by id's. These are created automatically if no was provided by the user.
Also I implemented Display for each Config type. This makes it way easier to print out current configurations.
Also broken configurations aren't lost from now on. If the user breaks the configuration by editing the config file, the old content is saved in a separate file. He then has the option to load some parts of it (TODO).

Next step is to parse the start subcommand to give control back to the main function.