created a template algorithm file & added the start subcommand
The template algorithm file should be used in the future to write algorithms. Also I'll use it to program the right interface and provide the needed functionality.

The start subcommand is not working yet. But it compiles and I will add the functionality for it in the next few commits.