created algorithm-utils
algorithm-utils is a crate meant to load algorithms dynamically at run time. This provides the option to add pre compiled algorithms without needing to recompile the whole application. (This could be especially interesting for non programmers)

Also algorithm-utils will hold all the fundamental structs and enums needed for trading, like an Order or a Position.

Currently algorithm-utils is not finished, yet.  But it compiles and loading algorithms already worked in a little test crate.