enabled the ability to load algorithms, improved some modules in parse_args & made some minor changes in order.rs
It's now possible to load external algorithms!!! Finally! And it's also possible to call functions defined in these algorithms. This means from now on I can load algorithms to trade with!

Also I changed a view things in the argument parser. It now has the ability to manipulate algorithm related settings. There's also a Settings struct now. It provided the ability to hold the algorithms in the settings. This was not possible with ConfigFile since Algorithm does not implement some important traits.

The last thing I updated is the way how one cancel the other orders are saved. It will from now on be a linked list. This may be bad for performance but it's still the best way of doing it.