Made some minor changes in the initial argument parsing & moved the algorithm-utils into another directory
Moving the algorithm-utils into a separate directory and creating a new repository for this crate was a design decision made so both crates can improve independent of each other. Also this will make work much easier for future algorithm developers since they can just install the algorithm-utils crate instead of the whole application.

Next steps are to start with the comdirect api. For this I'll first need to implement a first version of the BrokerInterface (in algorithm-utils). Since I already experimented with the comdirect api in python it should be quite easy to implement the first version of the api. After that potentially the first algorithms could start paper trading.